% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_missing.R
\name{fix_missing}
\alias{fix_missing}
\title{Fix missing values}
\usage{
fix_missing(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
The processed \code{data.frame}.
}
\description{
Replace NAs in a data frame with the median for the column, and add a new
booolean column indicating which rows have had NAs replaced.
}
