% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericalise.R
\name{numericalise}
\alias{numericalise}
\title{Convert a non-numeric column to a numeric representation}
\usage{
numericalise(df, col, new_col_name = NULL, max_n_cats = Inf)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{col}{The name (as a string) of the column to be converted}

\item{new_col_name}{The name of the new column to be added to the df. If not
supplied, defaults to "[col]_numeric".}

\item{max_n_cats}{Columns containing greater than this number of unique
values will not be numericalised. If this is not supplied, the column will
be converted regardless.}
}
\value{
The input \code{data.frame} with an addtional column containing the
  numeric representation of \code{col}.
}
\description{
Given an input \code{data.frame}, the specified column will be converted to
numeric values unless the number of unique values is greater than
\code{max_n_cats}. This is a wrapper for \code{as.numeric(as.factor(col))}.
}
