% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encode.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-hot encode a variable}
\usage{
one_hot_encode(df, col, max_n_cats = Inf, drop_one_col = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{col}{The column to be encoded.}

\item{max_n_cats}{Columns will only be one-hot encoded if they contain this
many categories or fewer.}
}
\value{
A \code{data.frame}.
}
\description{
Given an input \code{data.frame} and a target column, one-hot encode the
column if the number of unique values is greater than a specified threshold.
The original column is dropped. To prevent whitespace appearing in column
names, any \code{[:space:]} characters are replaced with underscores using
\code{gsub}.
}
