% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_df.R
\name{proc_df}
\alias{proc_df}
\title{Process a data.frame}
\usage{
proc_df(df, y = NULL, do_scale = FALSE, nas = NULL,
  preproc_fun = NULL, max_n_cats = Inf, subset = NULL,
  mapper = NULL)
}
\arguments{
\item{df}{The data frame of interest.}

\item{y}{The dependent variable. If not specified it is assumed that
\code{df} contains only independent variables.}

\item{do_scale}{Boolean - should columns be scaled and centred?}

\item{preproc_fun}{A function to be applied to \code{df}. This will be simply
called as \code{df <- preproc_fn(df)}, so make sure it is fully
self-contained and returns a \code{data.frame}.}

\item{max_n_cats}{Column with less than or equal to this number of unique
entries will be converted to one-hot encoded columns.}

\item{subset}{If set, a random sample of \code{n = subset} rows will be taken.}

\item{mapper}{If \code{do_scale} is true, the mapper function calculates the
scalng params (mean and standard deviation) to be applied.}

\item{skip_fields}{A character vector of column names to be dropped from
\code{df}.}

\item{ignore_fields}{A character vector of column names to be ignored during
processing.}

\item{na_list}{List of NA column to add. NA columns are also added if there
are any mising values.}
}
\value{
A \code{list} containing three items: \code{x}, \code{y}, \code{NAs},
  and (optionally) \code{mapper}, a data.frame containing the mean and
  standard deviation of each column.
}
\description{
Port of the Python \code{fast.ai}'s \code{prod_df} function.
}
\details{
Given an input data.frame, split off the response variable and convert it
to an entirely numeric data.frame. NaNs are replaced using median imputation,
and a new boolean `[col]_na` column added to indicate where the missing
values were.
}
