% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vars.R
\name{scale_vars}
\alias{scale_vars}
\title{Transform numeric columns}
\usage{
scale_vars(df)
}
\arguments{
\item{df}{A `data.frame` to be transformed.}
}
\value{
A transformed `data.frame`.
}
\description{
If the `mapper` argument is supplied, equivalent to calling `mapper(df)`. If
mapper is not supplied then any numeric columns are scaled and centred to
have a mean of 0 and a standard deviation of 1.
}
