% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericalise.R
\name{numericalise}
\alias{numericalise}
\title{Convert a non-numeric column to a numeric representation}
\usage{
numericalise(df, col)
}
\arguments{
\item{df}{A \code{data.frame}}

\item{col}{The name (as a string) of the column to be converted}
}
\value{
The transformed \code{data.frame}.
}
\description{
Given an input \code{data.frame}, the specified column will be converted to
numeric values. This is a wrapper for \code{as.numeric(as.factor(df$col))},
or as.numeric(cdf$col) for boolean columns.
}
