% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_df.R
\name{proc_df}
\alias{proc_df}
\title{Process a data.frame}
\usage{
proc_df(df, y = NULL, do_scale = FALSE, normalise = FALSE,
  detect_outliers = FALSE, outlier_threshold = 1.5,
  clip_outliers = FALSE, flag_outliers_both_directions = FALSE,
  nas = NULL, preproc_fun = NULL, max_n_cats = Inf, subset = NULL,
  mapper = NULL)
}
\arguments{
\item{df}{The data frame of interest.}

\item{y}{The dependent variable. If not specified it is assumed that
\code{df} contains only independent variables.}

\item{do_scale}{Should columns be scaled and centred?}

\item{normalise}{See \code{\link{scale_vars}}.}

\item{detect_outliers}{Should outliers be flagged? If \code{TRUE}, also set
\code{outlier_threshold}, \code{clip_outliers} and
\code{flag_outliers_both_directions}. See \link{\code{detect_outliers}}.}

\item{outlier_threshold}{See \code{\link{detect_outliers}}.}

\item{clip_outliers}{See \code{\link{detect_outliers}}.}

\item{flag_outliers_both_directions}{See \code{\link{detect_outliers}}.}

\item{preproc_fun}{A function to be applied to \code{df}. This will be simply
called as \code{df <- preproc_fn(df)} before any other operations, so make
sure it is fully self-contained and returns a \code{data.frame}.}

\item{max_n_cats}{Column with less than or equal to this number of unique
entries will be converted to one-hot encoded columns.}

\item{subset}{If set, a random sample of \code{subset} rows will be taken.}

\item{mapper}{If \code{do_scale} is true, the mapper function calculates the
scaling params (mean and standard deviation) to be applied.}

\item{skip_fields}{A character vector of column names to be dropped from
\code{df}. Not yet implemented.}

\item{ignore_fields}{A character vector of column names to be ignored during
processing. Not yet implemented.}

\item{na_list}{List of NA column to add. NA columns are also added if there
are any mising values. Not yet implemented.}
}
\value{
A transformed \code{data.frame}.
}
\description{
Port of \href{https://github.com/fastai/fastai}{fast.ai}'s \code{proc_df} function.
}
\details{
Given an input data.frame, split off the response variable and convert it
to an entirely numeric data.frame. NaNs are replaced using median imputation,
and a new boolean `[col]_na` column added to indicate where the missing
values were.
}
